// alyusr.client/alyusr.client/ClientApp/src/models/company/companySetting.ts
// export interface CompanySetting {
//   Errors: any[];
//   Setting: SettingDetails;
//   Status: number;
// }

export interface CompanySetting {
  FormatDecimal: string;
  DefaultCurrency_ID: number;
  // currency: Currency
  Currency_ID: number;
  CurrencyShortCut: string;
  IsTrialVersion: boolean;
  SoftwareName: any;
  Advert: any;
  PrintItemPrescriptionOfTransaction: boolean;
  BacckGroundImage: string;
  LogoImage: string;
  BackColor: number;
  BackGroundImageName: string;
  LogoImageName: string;
  Phone: string;
  Mail: string;
  Password: string;
  DateType: boolean;
  TextLogo: string;
  ValueOfPoint: number;
  SendMailOfSalesOnClose: boolean;
  RecivedMail: string;
  Header: any;
  DecimalPlace: number;
  PathBackUp: string;
  UserNameSMS: string;
  PasswordSMS: string;
  MobileReceiver: string;
  SendSMS: boolean;
  SenderName: string;
  SetScaleOnPrice: boolean;
  SetScaleWeightonKG: boolean;
  UseOldModelForSearch: boolean;
  DefaultTaxPercentage: number;
  ApplyTax: boolean;
  TaxNumber: string;
  HideItemFromCloseDay: boolean;
  MaxQuantity: number;
  TransferTaxToDiscount: boolean;
  ShowItemPriceWithoutTax: boolean;
  PrintItemArabicEngInBill: boolean;
  Company_Address: string;
  CalcDiscountWithVAT: boolean;
  SetCostPriceZeroOnSalePriceZero: boolean;
  Is_Company_Authorized_Tobacco: boolean;
  WaterMarkImage: string;
  ServiceUI: string;
  IsUploadTransactionToCloud: boolean;
  Printer_Tablet_List: string;
  SystemSettings: string;
  IncludeVatOnEditPrice: boolean;
  DaysOfDeleteBackupFiles: number;
  IsCodeTransactionSeparated: boolean;
  CompanyAddress: string;
  IsSentEmailReportAfterCloseDayOnly: boolean;
  CompanySetting_UniqueId: string;
  TermsAndCondations: string;
  IsFloorSecondeCurrency: boolean;
  CheckBalanceOfItemOnAdd: boolean;
  LogoImage64: string;
  ValueOfCurrency: number;
  FormatDate: string;
  ArabicLanguage: string;
  EnglishLanguage: string;
  voucherSettings: any;
  AttachmentFolder: any;
  IsUploadShipmentTrackToCloud: boolean;
  Token: any;
  IsEnableToUploadEinvoice: boolean;
  ID: number;
  CreatedBy: any;
  ModifiedBy: any;
  Name: string;
  CreationDate: string;
  ModificationDate: string;
  VerifyOnUpdate: boolean;
  Errors: any[];
  rowState: number;
}